extract the core of the configurable logger from access_log.c so that it could be used by the status handler
