+++ CMakeLists.txt
@@ -214,6 +214,7 @@ SET(LIB_SOURCE_FILES
     lib/core/configurator.c
     lib/core/context.c
     lib/core/headers.c
+    lib/core/logconf.c
     lib/core/proxy.c
     lib/core/request.c
     lib/core/token.c
